+++ src/design-json-property.js
@@ -34,6 +34,10 @@ export default class DesignJsonProperty {
   /**
    * @type {string}
    */
+  static STUDIO_LINK_ENABLED = 'studioLinkEnabled';
+  /**
+   * @type {string}
+   */
   static DEFAULT_LOCALE = 'defaultLocale';
   /**
    * @type {string}